% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allGenotypes}
\alias{allGenotypes}
\title{Genotype matrix}
\usage{
allGenotypes(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
An integer matrix with two columns and \code{choose(n+1, 2)} rows.
}
\description{
An autosomal marker with \code{n} alleles has \code{choose(n+1, 2)} possible
unordered genotypes. This function returns these as rows in a
matrix.
}
\examples{
allGenotypes(3)

}
