% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.grid.subset.R
\name{geno.grid.subset}
\alias{geno.grid.subset}
\title{Genotype combinations}
\usage{
geno.grid.subset(x, partialmarker, ids, make.grid = T)
}
\arguments{
\item{x}{a \code{\link[=ped]{ped()}} object.}

\item{partialmarker}{a \code{\link[=marker]{marker()}} object compatible with \code{x}.}

\item{ids}{a vector with ID labels of one or more pedigree members.}

\item{make.grid}{a logical. If FALSE, a list is returned, otherwise
\code{fast.grid} is applied to the list before returning it.}
}
\description{
Auxiliary functions computing possible genotype combinations in a pedigree.
These are not normally intended for end users.
}
