% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{likelihood_merlin}
\alias{likelihood_merlin}
\title{Pedigree likelihood computed by MERLIN}
\usage{
likelihood_merlin(x, markers = seq_len(nMarkers(x)), logbase = NULL,
  verbose = FALSE, generate.files = TRUE, cleanup = generate.files,
  logfile = "")
}
\arguments{
\item{x}{a \code{\link{ped}} object}

\item{markers}{an integer vector indicating which markers to use (default:
all).}

\item{logbase}{a positive number, or NULL. If numeric, the log-likelihood is
returned, with \code{logbase} as basis for the logarithm.}

\item{verbose}{a logical: Show MERLIN output and other information, or not.}

\item{generate.files}{a logical. If TRUE, the files 'merlin.ped',
'merlin.dat', 'merlin.map', and 'merlin.freq'.}

\item{cleanup}{a logical: Should the MERLIN files be deleted automatically?}

\item{logfile}{a character. If this is given, the MERLIN screen output will
be written to a file with this name.}
}
\value{
A number.
}
\description{
For this function to work, MERLIN must be installed and the path to
merlin.exe included in the PATH variable. The \code{likelihood_merlin} function is
a wrapper for one particular functionality of MERLIN.
}
\details{
By default the following MERLIN command is run via \code{\link[=system]{system()}},
after creating appropriate files in the current working directory:

\preformatted{%
  merlin -p _merlin.ped -d _merlin.dat -m _merlin.map -f _merlin.freq
         --lik --bits:100 --megabytes:4000 --quiet
}
}
\examples{

\dontrun{
x = pedtools::nuclearPed(1)
m = pedtools::marker(x, "3"=1:2)
x = pedtools::setMarkers(x, m)

# Likelihood computation by MERLIN:
likelihood_merlin(x)
}

}
\references{
\url{http://csg.sph.umich.edu/abecasis/Merlin/}
}
\author{
Magnus Dehli Vigeland
}
