% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedprobr-package.R
\docType{package}
\name{pedprobr-package}
\alias{pedprobr}
\alias{pedprobr-package}
\title{pedprobr: Probability Computations on Pedigrees}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of the Elston-Stewart algorithm for calculating pedigree likelihoods given genetic marker data (Elston and Stewart (1971) \doi{10.1159/000152448}). The standard algorithm is extended to allow inbred founders. 'pedprobr' is part of the 'pedsuite', a collection of packages for pedigree analysis in R. In particular, 'pedprobr' depends on 'pedtools' for pedigree manipulations and 'pedmut' for mutation modelling. For more information, see 'Pedigree Analysis in R' (Vigeland, 2021, ISBN:9780128244302).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/pedprobr}
  \item Report bugs at \url{https://github.com/magnusdv/pedprobr/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
