% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedprobr-package.R
\docType{package}
\name{pedprobr}
\alias{pedprobr}
\alias{pedprobr-package}
\title{pedprobr: Pedigree probabilities in R}
\description{
pedprobr implements the Elston-Stewart algorithm for computing probabilities
of marker genotypes on pedigrees. It build on the pedtools package which
handles pedigrees and marker objects.
}
