% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoMarkerDistribution.R
\name{twoMarkerDistribution}
\alias{twoMarkerDistribution}
\title{Probability distribution for linked marker}
\usage{
twoMarkerDistribution(x, id, partialmarker1, partialmarker2, theta,
  loop_breakers = NULL, eliminate = 99, verbose = TRUE)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{id}{A single ID label.}

\item{partialmarker1, partialmarker2}{Either a \code{marker} object or the name (or
index) of a marker attached to \code{x}.}

\item{theta}{A single numeric in the interval \code{[0, 0.5]}: the recombination
fraction between the two markers.}

\item{loop_breakers}{(Only relevant if the pedigree has loops). A vector with
ID labels of individuals to be used as loop breakers. If NULL (default)
loop breakers are selected automatically. See \code{\link[=breakLoops]{breakLoops()}}.}

\item{eliminate}{A non-negative integer, indicating the number of iterations
in the internal algorithm for reducing the genotype space. Positive values
can save time if \code{partialmarker1} and/or \code{partialmarker2} have many
alleles.}

\item{verbose}{A logical.}
}
\value{
A named matrix giving the joint genotype distribution.
}
\description{
Computes the joint genotype distribution of two markers for a specified
pedigree member, conditional on known genotypes and the recombination rate
between the markers.
}
\examples{

library(pedtools)

x = nuclearPed(father="fa", mother="mo", children=c("ch1", "ch2"))

# Empty SNP marker
emptySNP = marker(x, alleles=1:2)
SNP1 = marker(x, fa=c(1,1), mo=c(1,0), alleles=1:2, afreq=c(0.1, 0.9))
twoMarkerDistribution(x, id="mo", emptySNP, SNP1, theta=0)
twoMarkerDistribution(x, id="mo", emptySNP, SNP1, theta=0.5)
twoMarkerDistribution(x, id="ch1", emptySNP, SNP1, theta=0.5)

# X-linked example
SNPX_1 = marker(x, mo=c('a','b'), ch1='b', alleles=c('a','b'), chrom=23)
SNPX_2 = marker(x, mo=c('a','b'), ch1='b', alleles=c('a','b'), chrom=23)
r1 = twoMarkerDistribution(x, id="ch2", SNPX_1, SNPX_2, theta=0)
r2 = twoMarkerDistribution(x, id="ch2", SNPX_1, SNPX_2, theta=0.5)
stopifnot(all(r1==c(.5,0,0,.5)), all(r2==c(.25,.25,.25,.25)))

}
\seealso{
\code{\link[=oneMarkerDistribution]{oneMarkerDistribution()}}
}
\author{
Magnus Dehli Vigeland
}
