% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allGenotypes}
\alias{allGenotypes}
\title{Genotype matrix}
\usage{
allGenotypes(n)
}
\arguments{
\item{n}{A positive integer.}
}
\description{
An autosomal marker with \code{n} alleles has \code{G = choose(n+1, 2)} possible
genotypes (when the allele order within a genotype does not matter). This
function returns these in the form of a \code{G * 2} matrix where each row is a
genotype. The computation is trivial; the main point is to provide a
consistent ordering of the genotypes. The homozygous genotypes come first,
followed by the heterozygous.
}
\examples{
allGenotypes(3)

}
