% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedprobr-package.R
\docType{package}
\name{pedprobr}
\alias{pedprobr}
\alias{pedprobr-package}
\title{pedprobr: Probability Computations on Pedigrees}
\description{
An implementation of the Elston-Stewart algorithm for calculating pedigree
likelihoods given genetic marker data (Elston and Stewart (1971),
\doi{10.1159/000152448}). The standard algorithm is extended to allow inbred
founders. Mutation modelling is included via the 'pedmut' package. 'pedprobr'
is part of the ped suite, a collection of packages for pedigree analysis in
R, based on 'pedtools' for handling pedigrees and markers.
}
